added ability to add and remove items from inventory

+++ classes/needles.cs
@@ -1,9 +1,7 @@
-using System;
-
 public class Needles
 {
   public int id { get; set; } 
-  public int size { get; set; } 
+  public double size { get; set; } 
   public string type { get; set; } 
   public string material { get; set; } 
 }
